feat(cli/wizard): move activate command from user to new wizard subcommand
The activate flow performs full system activation (vault + system + network +
password reset), not just user-scoped operations. Promote it to a dedicated
top-level `wizard activate` subcommand for clearer semantics, and remove the
old `user activate` entry. Behavior, flags, and validation are unchanged.

Made-with: Cursor
