+++ cli/cmd/ctl/root.go
@@ -12,6 +12,7 @@ import (
 	"github.com/beclab/Olares/cli/cmd/ctl/os"
 	"github.com/beclab/Olares/cli/cmd/ctl/osinfo"
 	"github.com/beclab/Olares/cli/cmd/ctl/user"
+	"github.com/beclab/Olares/cli/cmd/ctl/wizard"
 	"github.com/beclab/Olares/cli/version"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -49,6 +50,7 @@ func NewDefaultCommand() *cobra.Command {
 	cmds.AddCommand(gpu.NewCmdGpu())
 	cmds.AddCommand(amdgpu.NewCmdAmdGpu())
 	cmds.AddCommand(user.NewUserCommand())
+	cmds.AddCommand(wizard.NewWizardCommand())
 	cmds.AddCommand(disk.NewDiskCommand())
 	cmds.AddCommand(app.NewAppCommand())
