+++ cli/cmd/ctl/user/root.go
@@ -11,7 +11,6 @@ func NewUserCommand() *cobra.Command {
 	cmd.AddCommand(NewCmdDeleteUser())
 	cmd.AddCommand(NewCmdListUsers())
 	cmd.AddCommand(NewCmdGetUser())
-	cmd.AddCommand(NewCmdActivateUser())
 	cmd.AddCommand(NewCmdResetPassword())
 	// cmd.AddCommand(NewCmdUpdateUserLimits())
 	return cmd
